backend/python/requirements.txt
grpcio
grpcio-tools
protobuf
psycopg2-binary 